+++ modules/core/src/decorator/serialize.decorator.ts
@@ -1,17 +1,26 @@
-declare var Reflect:any;
+declare var Reflect: any;
 
 export const METADATA_KEY_SERIALIZABLE = "SERIALIZABLE";
 
-export function serializable(context, key) {
+export interface SerializableProperty {
 
-    var serializable = Reflect.getMetadata(METADATA_KEY_SERIALIZABLE, context) || [];
-
-    serializable.push(key);
-
-    Reflect.defineMetadata(METADATA_KEY_SERIALIZABLE, serializable, context);
+    key: string;
+    name: string;
 }
 
-export function getSerializable (context): Array<string> {
+export function serializable(name?: string) {
+
+    return function (context, key) {
+
+        var serializable = Reflect.getMetadata(METADATA_KEY_SERIALIZABLE, context) || [];
+
+        serializable.push({key: key, name: name || key});
+
+        Reflect.defineMetadata(METADATA_KEY_SERIALIZABLE, serializable, context);
+    };
+}
+
+export function getSerializable(context): Array<SerializableProperty> {
 
     return Reflect.getMetadata(METADATA_KEY_SERIALIZABLE, context) || [];
 }