+++ modules/core/src/model/checkbox/dynamic-checkbox-group.model.ts
@@ -7,9 +7,9 @@ export const DYNAMIC_FORM_CONTROL_TYPE_CHECKBOX_GROUP = "CHECKBOX_GROUP";
 
 export class DynamicCheckboxGroupModel extends DynamicFormGroupModel {
 
-    @serializable group: Array<DynamicCheckboxModel>;
+    @serializable() group: Array<DynamicCheckboxModel>;
 
-    @serializable readonly type: string = DYNAMIC_FORM_CONTROL_TYPE_CHECKBOX_GROUP;
+    @serializable() readonly type: string = DYNAMIC_FORM_CONTROL_TYPE_CHECKBOX_GROUP;
 
     constructor(config: DynamicFormGroupModelConfig, cls?: ClsConfig) {