+++ modules/core/src/model/form-group/dynamic-form-group.model.ts
@@ -21,11 +21,11 @@ export interface DynamicFormGroupModelConfig extends DynamicFormControlModelConf
 export class DynamicFormGroupModel extends DynamicFormControlModel implements DynamicFieldSet {
 
     asyncValidator: AsyncValidatorFn | null;
-    @serializable group: Array<DynamicFormControlModel> = [];
-    @serializable legend: string | null;
+    @serializable() group: Array<DynamicFormControlModel> = [];
+    @serializable() legend: string | null;
     validator: ValidatorFn | null;
 
-    @serializable readonly type: string = DYNAMIC_FORM_CONTROL_TYPE_GROUP;
+    @serializable() readonly type: string = DYNAMIC_FORM_CONTROL_TYPE_GROUP;
 
     constructor(config: DynamicFormGroupModelConfig, cls?: ClsConfig) {