+++ modules/core/src/model/select/dynamic-select.model.ts
@@ -6,7 +6,7 @@ export const DYNAMIC_FORM_CONTROL_TYPE_SELECT = "SELECT";
 
 export class DynamicSelectModel<T> extends DynamicOptionControlModel<T> {
 
-    @serializable readonly type: string = DYNAMIC_FORM_CONTROL_TYPE_SELECT;
+    @serializable() readonly type: string = DYNAMIC_FORM_CONTROL_TYPE_SELECT;
 
     constructor(config: DynamicOptionControlModelConfig, cls?: ClsConfig) {