+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### **Breaking Changes** 
 
 * **!!! IMPORTANT !!!**: `[(ngModel)]` bindings have been removed. If you need to update the `value` of any `DynamicFormControlModel` 
-at run time use `valueChanges` `BehaviorSubject` [(see README.md)](https://github.com/udos86/ng2-dynamic-forms#model-bindings-and-control-references) 
+at run time use `valueUpdates` `Subject` [(see README.md)](https://github.com/udos86/ng2-dynamic-forms#model-bindings-and-control-references) 
 
 ### **Bugfixes**